Move create_dd.h to python_bindings
It's clearly DD-specific and is not used anywhere but in Python
bindings.
