+++ src/python_bindings/dd/create_dd.cpp
@@ -1,4 +1,4 @@
-#include "core/util/create_dd.h"
+#include "python_bindings/dd/create_dd.h"
 
 namespace util::dd {
 model::DFStringConstraint CreateDf(std::string_view const& col_name, double const lower_bound,
